--- a/test/references/pluto01_5_crc32c.txt
+++ b/test/references/pluto01_5_crc32c.txt
@@ -1,0 +1,1 @@
+d6eb83dbaa46be3c21f164ea6fcd7ee41c6cc3efdabadb783e4bce31bfb0f04f