--- a/test/references/ref07_5_sha1.txt
+++ b/test/references/ref07_5_sha1.txt
@@ -1,0 +1,1 @@
+86b9ab4a8278faf79f6b67dedd8fa311aeffafeb