--- a/test/references/ref13_5_sha1.txt
+++ b/test/references/ref13_5_sha1.txt
@@ -1,0 +1,1 @@
+e7eade7f5f6a1b4618bc72c7cc8cd99d608ed67a