--- a/test/references/ref16_5_sha1.txt
+++ b/test/references/ref16_5_sha1.txt
@@ -1,0 +1,1 @@
+f79e4fc3bb0aeb0cada712c074beb8700cee36d6