--- a/test/references/ref18_5_sha1.txt
+++ b/test/references/ref18_5_sha1.txt
@@ -1,0 +1,1 @@
+de5b8d8cf2cd70b0e6222934fa7bcc4191ca2cd4