--- a/test/references/ref22_5_sha256.txt
+++ b/test/references/ref22_5_sha256.txt
@@ -1,0 +1,1 @@
+4d876c1daadb8a5caf8fcd8106356db7d52183db3a3decbedfe46f433bf6d5de