## Description

There are several improvements slated for `HashVersion{5}`. All of these rely on the new package extensions possible starting in Julia 1.9. With StableHashTraits 2.0, we no longer are supporting julia versions lower than 1.10, and so the upgrade opens up several improvements I didn't want to make until the use of extensions was possible.

The first such improvement, the TimeZones PR, optimizes the hashing of `ZonedDateTime` objects, by computing the absolute date time, for UTC-0, thereby bypassing a large object tracking time zone history referenced by all `ZonedDatTime` objects.

## Benchmarks

Author, please:

**TODO**

Run the benchmarks under `test/benchmarks.jl` locally, and update `test/benchmark_records.md` by inserting a new table at the end of the document with the updated numbers
printed at the end of the benchmark script. Place the table as it occurs before
and after the application of this PR.

### Before

```
16×6 DataFrame
 Row │ benchmark   hash       version    base        trait       ratio
     │ SubStrin…   SubStrin…  SubStrin…  String      String      Float64
─────┼─────────────────────────────────────────────────────────────────────
   1 │ dicts       crc        4          1.533 ms    88.197 ms   57.5258
   2 │ structs     crc        4          57.417 μs   500.375 μs   8.71475
   3 │ tuples      crc        4          57.542 μs   344.500 μs   5.98693
   4 │ numbers     crc        4          28.875 μs   122.042 μs   4.22656
   5 │ dataframes  crc        4          73.042 μs   277.958 μs   3.80545
   6 │ symbols     crc        4          1.155 ms    1.371 ms     1.18639
   7 │ missings    crc        4          240.917 μs  207.083 μs   0.859562
   8 │ strings     crc        4          879.208 μs  289.083 μs   0.328799
   9 │ dicts       sha256     4          1.948 ms    81.803 ms   41.9871
  10 │ tuples      sha256     4          484.250 μs  1.599 ms     3.30201
  11 │ structs     sha256     4          475.292 μs  1.348 ms     2.83615
  12 │ symbols     sha256     4          2.009 ms    3.023 ms     1.50492
  13 │ dataframes  sha256     4          497.667 μs  702.000 μs   1.41058
  14 │ numbers     sha256     4          246.000 μs  336.750 μs   1.3689
  15 │ strings     sha256     4          1.753 ms    1.707 ms     0.973521
  16 │ missings    sha256     4          477.875 μs  443.042 μs   0.927109
```

### After

```
 
```