@haberdashPI 2024-11-27
Turns out we don't need this. I was getting confusingly good benchmarks for `version=4`. (see last comment in #26 for details)